backend/requirements.txt
Django==1.11.21
django-cors-headers==1.2.0
django-oauth-toolkit==0.10.0
djangorestframework==3.9.1
django-webpack-loader==0.6.0
pandas==0.25.2
psycopg2==2.8.4
pytest==5.2.2